introduce NeuralDataOriginator
- replace TDTOriginator and HTKOriginator
